switch sql logic to databaseMethods
